requirements.txt
plotly==5.23.0
seaborn
pytorch